Combine SettingsLinkedEmail panel and Setting LinkedLinkedPhone panel into SettingsLinkedAccountPanel [#1407]
